Rename tests to reflect on the function names
